move configuration delegation to main module.
